happy path for identifyRepository

+++ Tests/Shared/TrackEventData.swift
@@ -5,13 +5,13 @@ import Foundation
 public struct TrackEventData: Codable, Equatable {
     var product: String
     var count: Int
-    var price: Double
+    var price: Int
 }
 
 public extension TrackEventData {
     static func random(update: Bool = false) -> TrackEventData {
         TrackEventData(product: String.random,
                        count: Int.random(in: 1 ... 10),
-                       price: Double.random(in: 10 ... 100))
+                       price: Int.random(in: 10 ... 100))
     }
 }
